+++ src/Entities/Enums/LodgingItemType.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace GlobalPayments\Api\Entities\Enums;
+
+use GlobalPayments\Api\Entities\Enum;
+
+class LodgingItemType extends Enum
+{
+    const RESTAURANT = 'RESTAURANT';
+    const GIFT_SHOP = 'GIFT_SHOP';
+    const MINI_BAR = 'MINI_BAR';
+    const PHONE = 'PHONE';
+    const LAUNDRY = 'LAUNDRY';
+    const OTHER = 'OTHER';
+    const NO_SHOW = 'NO_SHOW';
+}