+++ src/Entities/LodgingItems.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace GlobalPayments\Api\Entities;
+
+class LodgingItems
+{
+    /** @var array */
+    public $types;
+
+    /** @var string */
+    public $reference;
+
+    /** @var string */
+    public $totalAmount;
+
+    /** @var array */
+    public $paymentMethodProgramCodes;
+}